split CUP patch
