--- addons/missileGuidanceCompat/CfgMagazines.hpp
@@ -1,7 +1,0 @@
-class CfgMagazines {
-    class CA_LauncherMagazine;
-    class CUP_Dragon_EP1_M: CA_LauncherMagazine {
-        ammo = "ace_dragon_super";
-        initSpeed = 120;
-    };
-};
