+++ addons/missileGuidanceCompat/CfgMissileTypesNato.hpp
@@ -1,0 +1,20 @@
+class GVAR(type_AMRAAM);
+class GVAR(type_ASRAAM);
+class GVAR(type_Dagr);
+class GVAR(type_Dragon);
+class GVAR(type_ESSM);
+class GVAR(type_Hellfire);
+class GVAR(type_Hot);
+class GVAR(type_Javelin);
+class GVAR(type_Jdam);
+class GVAR(type_Maverick);
+class GVAR(type_Milan);
+class GVAR(type_Nlaw);
+class GVAR(type_Patriot);
+class GVAR(type_Paveway);
+class GVAR(type_RAM);
+class GVAR(type_RBS70);
+class GVAR(type_Redeye);
+class GVAR(type_Sidewinder);
+class GVAR(type_Stinger);
+class GVAR(type_TOW);
