+++ addons/missileGuidanceCompat/CfgMissileTypesWarsaw.hpp
@@ -1,0 +1,25 @@
+class ACEGVAR(missileguidance,type_Ataka);
+class ACEGVAR(missileguidance,type_Bastion);
+class ACEGVAR(missileguidance,type_Drakon);
+class ACEGVAR(missileguidance,type_Fagot);
+class ACEGVAR(missileguidance,type_Falanga);
+class ACEGVAR(missileguidance,type_Fleyta);
+class ACEGVAR(missileguidance,type_Igla);
+class ACEGVAR(missileguidance,type_KAB);
+class ACEGVAR(missileguidance,type_KH25);
+class ACEGVAR(missileguidance,type_KH29);
+class ACEGVAR(missileguidance,type_Kobra);
+class ACEGVAR(missileguidance,type_Konkurs);
+class ACEGVAR(missileguidance,type_Kornet);
+class ACEGVAR(missileguidance,type_Malyutka);
+class ACEGVAR(missileguidance,type_Metis);
+class ACEGVAR(missileguidance,type_Molniya);
+class ACEGVAR(missileguidance,type_R73);
+class ACEGVAR(missileguidance,type_R74);
+class ACEGVAR(missileguidance,type_R77);
+class ACEGVAR(missileguidance,type_Refleks);
+class ACEGVAR(missileguidance,type_S400);
+class ACEGVAR(missileguidance,type_Shturm);
+class ACEGVAR(missileguidance,type_Strela);
+class ACEGVAR(missileguidance,type_Vikhr);
+class ACEGVAR(missileguidance,type_Vympel);
